📦 refactor(event-handler): migrate to ES Modules with named exports
