+++ src/event-handler/wrap-error-handler.js
@@ -1,11 +1,7 @@
-"use strict";
-
-module.exports = wrapErrorHandler;
-
 // Errors thrown or rejected Promises in "error" event handlers are not handled
 // as they are in the webhook event handlers. If errors occur, we log a
 // "Fatal: Error occured" message to stdout
-function wrapErrorHandler(handler, error) {
+export function wrapErrorHandler(handler, error) {
   let returnValue;
 
   try {
