+++ test/integration/event-handler-test.js
@@ -1,8 +1,7 @@
-const test = require("tap").test;
-
-const createEventHandler = require("../../pkg/dist-src/event-handler");
-const pushEventPayload = require("../fixtures/push-payload");
-const installationCreatedPayload = require("../fixtures/installation-created-payload");
+import { test } from "tap";
+import { createEventHandler } from "../../pkg/dist-src/event-handler";
+import pushEventPayload from "../fixtures/push-payload.json";
+import installationCreatedPayload from "../fixtures/installation-created-payload.json";
 
 test("events", (t) => {
   t.plan(6);
